+++ h2gis-functions/src/main/java/org/h2gis/functions/io/DriverManager.java
@@ -97,7 +97,8 @@ public class DriverManager extends AbstractFunction implements ScalarFunction, D
             if(driverDef.getFileExt().equalsIgnoreCase(ext)) {
                 try (Statement st = connection.createStatement()) {
                     String tableName_ = TableLocation.parse(tableName, dbType).toString();
-                    st.execute(String.format("CREATE TABLE %s COMMENT %s ENGINE %s WITH %s",
+                    st.execute(String.format(
+                            "CREATE TABLE %s COMMENT %s ENGINE %s WITH %s",
                             tableName_,StringUtils.quoteStringSQL(fileName),
                             StringUtils.quoteJavaString(driverDef.getClassName()),StringUtils.quoteJavaString(fileName)));
                      return new String[]{tableName_};