+++ docs/CHANGELOG.md
@@ -14,3 +14,4 @@
 - Preserve the SRID when a “SELECT ST_GEOMFROMTEXT(''POINT(0 0)'', 4326) As the_geom” query is stored in an FGB table
 - Fix ST_SubDivide with empty geometry, improve performance
 - Add a new parameter, angle to ST_MakeGrid function
+- Fix ST_ClustersTest with String.format